% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACL-package.R
\docType{package}
\name{ACL-package}
\alias{ACL}
\alias{ACL-package}
\title{ACL: A Survey-Based Statistical Catch-at-Length Model}
\description{
Implements an age-length structured (ACL) operating model for survey-based statistical catch-at-length stock assessment. Designed for hard-to-age fisheries stocks, the package provides tools for model fitting, simulation testing, retrospective analysis, and diagnostic visualization.
}
\author{
\strong{Maintainer}: Hongyu Lin \email{linbojun99@gmail.com} (\href{https://orcid.org/0000-0001-6226-179X}{ORCID})

Authors:
\itemize{
  \item Fan Zhang \email{f-zhang@shou.edu.cn} (\href{https://orcid.org/0000-0003-0214-1790}{ORCID})
}

Other contributors:
\itemize{
  \item Sisong Dong [contributor]
}

}
\keyword{internal}
