% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bio_vars.R
\name{calculate_bio_vars}
\alias{calculate_bio_vars}
\title{Calculate biological and fishing variables}
\usage{
calculate_bio_vars(
  params,
  nage,
  rec.age,
  len_mid,
  len_border,
  M,
  init_Z,
  vbk,
  Linf,
  t0,
  mat_L50,
  mat_L95,
  cv_L,
  cv_inc,
  std_logR,
  std_logN0,
  alpha,
  beta,
  std_SN,
  q_surv_L50,
  q_surv_L95
)
}
\arguments{
\item{params}{A list containing the parameters variables.}

\item{nage}{Integer, number of ages.}

\item{rec.age}{Integer, the age of recruitment to the fishery.}

\item{len_mid}{Numeric vector, the midpoints of length bins.}

\item{len_border}{Numeric vector, the border values of length bins.}

\item{M}{Numeric, natural mortality.}

\item{init_Z}{Numeric, initial total mortality leading to equilibrium age structure.}

\item{vbk}{Numeric, von Bertalanffy growth coefficient.}

\item{Linf}{Numeric, asymptotic length in von Bertalanffy growth function.}

\item{t0}{Numeric, theoretical age at length zero in von Bertalanffy growth function.}

\item{mat_L50}{Numeric, length at which 50\% of individuals are mature.}

\item{mat_L95}{Numeric, length at which 95\% of individuals are mature.}

\item{cv_L}{Numeric, the coefficient of variance in length at age at the beginning year.}

\item{cv_inc}{Numeric, the coefficient of variance in growth increment.}

\item{std_logR}{Numeric, standard deviation of recruitment.}

\item{std_logN0}{Numeric, standard deviation of initial number at age.}

\item{alpha}{Numeric, alpha parameter in the stock-recruitment relationship.}

\item{beta}{Numeric, beta parameter in the stock-recruitment relationship.}

\item{std_SN}{Numeric, survey measurement error.}

\item{q_surv_L50}{Numeric, survey catchability at length for 50\% catchability.}

\item{q_surv_L95}{Numeric, survey catchability at length for 95\% catchability.}
}
\value{
A list containing the biological and fishing variables.
}
\description{
This function calculates and returns biological and fishing variables for the simulation model.
}
