% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_model.R
\name{diagnose_model}
\alias{diagnose_model}
\title{Diagnose model results}
\usage{
diagnose_model(model_result)
}
\arguments{
\item{model_result}{The model output. This should be a list that contains "bound_hit", "converge", and "final_outer_mgc" components.}
}
\value{
NULL. The function prints out the diagnostics and does not return anything.
}
\description{
This function receives a model output and performs various checks, including checking if the model has reached its boundaries, if it has converged, and printing the final likelihood estimate.
}
\examples{
\dontrun{
diagnose_model(model_result)
}
}
