% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acl_theme.R
\name{.acl_lab}
\alias{.acl_lab}
\title{Get an axis label from the theme}
\usage{
.acl_lab(axis = "x", key)
}
\arguments{
\item{axis}{Character. "x" or "y".}

\item{key}{Character. Label key (e.g. "year", "abundance").}
}
\value{
Character string.
}
\description{
Get an axis label from the theme
}
\keyword{internal}
