% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acl_theme.R
\name{.acl_scale_x}
\alias{.acl_scale_x}
\title{Build x-axis scale with breaks and expand}
\usage{
.acl_scale_x(x_breaks = NULL, n_breaks = 10)
}
\arguments{
\item{x_breaks}{Numeric vector or NULL (override). If NULL, reads from
global theme. If global is also NULL, uses \code{pretty_breaks(n)}.}

\item{n_breaks}{Integer. Fallback number of pretty breaks (default 10).}
}
\value{
A ggplot2 scale layer.
}
\description{
Build x-axis scale with breaks and expand
}
\keyword{internal}
