% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acl_theme.R
\name{.acl_title}
\alias{.acl_title}
\title{Get a title from the theme}
\usage{
.acl_title(key)
}
\arguments{
\item{key}{Character. Title key (e.g. "N", "SSB_se").}
}
\value{
Character string.
}
\description{
Get a title from the theme
}
\keyword{internal}
