% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_params.R
\name{initialize_params}
\alias{initialize_params}
\title{Initialize parameters for the fish simulation}
\usage{
initialize_params(
  nyear = 100,
  rec.age = 1,
  first.year = 2020,
  nage = 15,
  M = 0.2,
  init_Z = 0.5,
  vbk = 0.2,
  Linf = 60,
  t0 = 1/60,
  len_mid = seq(6, 50, 2),
  len_border = seq(5, 51, 2),
  a = exp(-12),
  b = 3,
  mat_L50 = 35,
  mat_L95 = 40,
  cv_L = 0.2,
  cv_inc = 0.2,
  std_logR = 0.3,
  std_logN0 = 0.2,
  alpha = 400,
  beta = 10,
  std_SN = 0.2,
  q_surv_L50 = 15,
  q_surv_L95 = 20
)
}
\arguments{
\item{nyear}{Integer, number of years (default: 100)}

\item{rec.age}{Integer, the age of recruitment to fishery (default: 1)}

\item{first.year}{Integer, the first year of projection (default: 2020)}

\item{nage}{Integer, number of ages (default: 15)}

\item{M}{Numeric, natural mortality (default: 0.2)}

\item{init_Z}{Numeric, initial total mortality leading to equilibrium age structure (default: 0.5)}

\item{vbk}{Numeric, von Bertalanffy growth parameter (default: 0.2)}

\item{Linf}{Numeric, asymptotic length (default: 60)}

\item{t0}{Numeric, growth parameter (default: 1/60)}

\item{len_mid}{Numeric vector, the midpoints of length bins.(default: seq(6,50,2))}

\item{len_border}{Numeric vector, the border values of length bins.(default: seq(5,51,2))}

\item{a}{parameter to define length–weight relationship.(default: exp(-12))}

\item{b}{parameter to define length–weight relationship.(default: 3)}

\item{mat_L50}{Numeric, length at which 50\% of individuals are mature.(default: 35)}

\item{mat_L95}{Numeric, length at which 95\% of individuals are mature.(default: 40)}

\item{cv_L}{Numeric, the coefficient of variance in length at age at the beginning year.(default: 0.2)}

\item{cv_inc}{Numeric, the coefficient of variance in growth increment.(default: 0.2)}

\item{std_logR}{Numeric, standard deviation of recruitment.(default: 0.3)}

\item{std_logN0}{Numeric, standard deviation of initial number at age.(default: 0.2)}

\item{alpha}{Numeric, Beverton-Holt model parameter.(default: 400)}

\item{beta}{Numeric, Beverton-Holt model parameter.(default: 10)}

\item{std_SN}{Numeric, survey measurement error.(default: 0.2)}

\item{q_surv_L50}{Numeric, survey catchability at length for 50\% catchability.(default: 15)}

\item{q_surv_L95}{Numeric, survey catchability at length for 95\% catchability.(default: 20)}
}
\value{
A list containing the initialized parameters and variables
}
\description{
This function initializes the parameters and variables needed for the fish simulation.(default:)
}
