% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_func.R
\name{mat_func}
\alias{mat_func}
\title{Maturation function}
\usage{
mat_func(L50, L95, length)
}
\arguments{
\item{L50}{Length at which 50\% of individuals are mature.}

\item{L95}{Length at which 95\% of individuals are mature.}

\item{length}{Length of the fish.}
}
\value{
Maturation probability.
}
\description{
Calculate maturation probability based on the length of the fish.
}
