% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BL.R
\name{plot_BL}
\alias{plot_BL}
\title{Plot BL over the years}
\usage{
plot_BL(
  model_result,
  line_size = 1.2,
  line_color = "black",
  line_type = "solid",
  facet_ncol = 3,
  facet_scales = "free"
)
}
\arguments{
\item{model_result}{A list that contains model output. The list should have a "report" component which contains a "BL" component representing counts in length groups.}

\item{line_size}{Numeric. The thickness of the line in the plot. Default is 1.5.}

\item{line_color}{Character. The color of the line in the plot. Default is "black".}

\item{line_type}{Character. The type of the line in the plot. Default is "solid".#' @return A ggplot object representing the plot.}

\item{facet_ncol}{Integer, number of columns in facet_wrap.}

\item{facet_scales}{Character, scales for facet_wrap.}
}
\description{
This function takes a list and a data frame as input, and plots BL (counts in length groups) over the years using ggplot2. The plot is faceted by the length groups.
}
\examples{
\dontrun{
plot_BL(model_result)
}
}
