% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CatL.R
\name{plot_CatL}
\alias{plot_CatL}
\title{Plot Estimated and Observed Catch-at-Length (Log or not) over the years or across lengths}
\usage{
plot_CatL(
  model_result,
  point_size = 1,
  point_color = "black",
  point_shape = 1,
  line_size = 1.2,
  line_color = "black",
  line_type = "solid",
  line_color1 = "red",
  line_color2 = "blue",
  facet_ncol = NULL,
  facet_scales = "free",
  type = c("length", "year"),
  exp_transform = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{model_result}{A list that contains the model output. The list should have a "report" component which contains an "Elog_index" component representing counts in length groups.}

\item{point_size}{Numeric. Specifies the size of the point in the plot. Default is 1.}

\item{point_color}{Character. Specifies the color of the point in the plot. Default is "black".}

\item{point_shape}{Numeric. Specifies the shape of the point in the plot. Default is 1.}

\item{line_size}{Numeric. Specifies the thickness of the line in the plot. Default is 1.2.}

\item{line_color}{Character. Specifies the color of the line in the plot. Default is "black".}

\item{line_type}{Character. Specifies the type of the line in the plot. Default is "solid".}

\item{line_color1}{Character. The color of the line for estimated abundance (Elog_index). Default is "red".}

\item{line_color2}{Character. The color of the line for observed abundance (logN_at_len). Default is "blue".}

\item{facet_ncol}{Integer. Specifies the number of columns in the facet_wrap. Default is NULL.}

\item{facet_scales}{Character. Specifies scales for facet_wrap. Default is "free".}

\item{type}{Character. It specifies whether the Elog_index is plotted across "length" or "year". Default is "length".}

\item{exp_transform}{Logical. Specifies whether to apply the exponential function to the data before plotting. If TRUE, the exponential of the data values is plotted. Default is FALSE.}

\item{return_data}{A logical indicating whether to return the processed data alongside the plot. Default is FALSE.}
}
\value{
A ggplot object representing the plot.
}
\description{
This function visualizes the Elog_index (counts in length groups) over years or across lengths using ggplot2.
The Elog_index represents the estimated abundance of the population within certain length groups, obtained from the model output.
The resulting plot is a line graph, with data points represented by distinct markers.
The function allows for faceted plots to better visualize the distribution of counts across different length groups or years.
}
\examples{
\dontrun{
# Use 'run_acl' to obtain 'model_result'
model_result <- run_acl(...)

# Plot Elog_index across lengths
plot_CatL(model_result, type = "length")

# Plot Elog_index over years
plot_CatL(model_result, type = "year")

# With exp_transform set to TRUE
plot_CatL(model_result, type = "...", exp_transform = TRUE)
}

}
