% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CNA.R
\name{plot_CNA}
\alias{plot_CNA}
\title{Plot catch numbers by age over the years}
\usage{
plot_CNA(
  model_result,
  line_size = 1,
  line_color = "red",
  line_type = "solid",
  facet_ncol = 3,
  facet_scales = "free"
)
}
\arguments{
\item{model_result}{A list that contains model output. The list should have a "report" component which contains a "BL" component representing counts in length groups.}

\item{line_size}{Line size for geom_line. Default is 1.}

\item{line_color}{Line color for geom_line. Default is "black".}

\item{line_type}{Line type for geom_line. Default is "solid".}

\item{facet_ncol}{Number of columns in facet wrap. Default is 3.}

\item{facet_scales}{Scales for facet wrap. Default is "free".}
}
\value{
A ggplot object representing the plot.
}
\description{
This function takes a matrix and a data frame as input, and plots fishing mortality over the years using ggplot2.
}
\examples{
\dontrun{
plot_CNA(a)
}
}
