% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NL.R
\name{plot_NL}
\alias{plot_NL}
\title{Plot NL over the years}
\usage{
plot_NL(
  model_result,
  line_size = 1.2,
  line_color = "red",
  line_type = "solid",
  facet_ncol = 3,
  facet_scales = "free",
  se = FALSE,
  se_color = "red",
  se_alpha = 0.2
)
}
\arguments{
\item{model_result}{A list that contains model output. The list should have a "report" component which contains a "NL" component representing counts in length groups.}

\item{line_size}{Numeric. The thickness of the line in the plot. Default is 1.5.}

\item{line_color}{Character. The color of the line in the plot. Default is "black".}

\item{line_type}{Character. The type of the line in the plot. Default is "solid".#' @return A ggplot object representing the plot.}

\item{facet_ncol}{Number of columns in facet wrap. Default is 3.}

\item{facet_scales}{Scales for facet wrap. Default is "free".}

\item{se}{Logical. Whether to calculate and plot standard error as confidence intervals. Default is FALSE.}

\item{se_color}{Character. The color of the confidence interval ribbon. Default is "blue".}

\item{se_alpha}{Numeric. The transparency of the confidence interval ribbon. Default is 0.2.}
}
\description{
This function takes a list and a data frame as input, and plots NL (counts in length groups) over the years using ggplot2. The plot is faceted by the length groups.
}
\examples{
\dontrun{
plot_NL(model_result)
}
}
