% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_number_age.R
\name{plot_number_age}
\alias{plot_number_age}
\title{Plot Number of different age groups per year}
\usage{
plot_number_age(
  model_result,
  line_size = 1,
  line_color = "red",
  line_type = "solid",
  facet_ncol = 3,
  facet_scales = "free",
  se = FALSE,
  se_color = "red",
  se_alpha = 0.2
)
}
\arguments{
\item{model_result}{A list that contains model output. The list should have a "report" component which contains a "na" component representing Number of different age groups per year.}

\item{line_size}{Line size for geom_line. Default is 1.}

\item{line_color}{Line color for geom_line. Default is "black".}

\item{line_type}{Line type for geom_line. Default is "solid".}

\item{facet_ncol}{Number of columns in facet wrap. Default is 3.}

\item{facet_scales}{Scales for facet wrap. Default is "free".}

\item{se}{Logical. Whether to calculate and plot standard error as confidence intervals. Default is FALSE.}

\item{se_color}{Character. The color of the confidence interval ribbon. Default is "blue".}

\item{se_alpha}{Numeric. The transparency of the confidence interval ribbon. Default is 0.2.}
}
\value{
A ggplot object representing the plot.
}
\description{
This function takes a matrix and a data frame as input, and plots na over the years using ggplot2.
}
\examples{
\dontrun{
plot_number_age(na, data.cl)
}
}
