% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OP.R
\name{plot_OP}
\alias{plot_OP}
\title{Plot OP over the years}
\usage{
plot_OP(
  model_result,
  data.CatL,
  point_size = 1,
  point_color = "black",
  point_shape = 1,
  line_size = 1.2,
  line_color = "black",
  line_type = "solid",
  facet_ncol = 3,
  facet_scales = "free"
)
}
\arguments{
\item{model_result}{A list that contains model output. The list should have a "report" component which contains a "Elog_index" component representing counts in length groups.}

\item{data.CatL}{A data frame that contains a "Year" column and multiple length group columns.}

\item{point_size}{Character. The size of the point in the plot. Default is "solid".#' @return A ggplot object representing the plot.}

\item{point_color}{Character. The color of the point in the plot. Default is "black".}

\item{point_shape}{Character. The shape of the point in the plot. Default is "solid".#' @return A ggplot object representing the plot.}

\item{line_size}{Numeric. The thickness of the line in the plot. Default is 1.5.}

\item{line_color}{Character. The color of the line in the plot. Default is "black".}

\item{line_type}{Character. The type of the line in the plot. Default is "solid".#' @return A ggplot object representing the plot.}

\item{facet_ncol}{Integer, number of columns in facet_wrap.}

\item{facet_scales}{Character, scales for facet_wrap.}
}
\description{
This function takes a list and a data frame as input, and plots Elog_index (counts in length groups) over the years using ggplot2. The plot is faceted by the length groups.
}
\examples{
\dontrun{
plot_OP(model_result, data.CatL)
}
}
