% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pla.R
\name{plot_pla}
\alias{plot_pla}
\title{Plot probability length at age}
\usage{
plot_pla(model_result, low_col = "white", high_col = "red")
}
\arguments{
\item{model_result}{A list that contains model output.}

\item{low_col}{color for low probabilities}

\item{high_col}{color for high probabilities}
}
\value{
ggplot object
}
\description{
Plot probability length at age
}
\examples{
\dontrun{
plot_prob_heatmap(model_result, "white", "steelblue")
}
}
