% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SBL.R
\name{plot_SBL}
\alias{plot_SBL}
\title{Plot SBL over the years}
\usage{
plot_SBL(
  model_result,
  data.CatL,
  line_size = 1.2,
  line_color = "SBLack",
  line_type = "solid",
  n_col = 3,
  scales = "free"
)
}
\arguments{
\item{model_result}{A list that contains model output. The list should have a "report" component which contains a "SBL" component representing counts in length groups.}

\item{data.CatL}{A data frame that contains a "Year" column and multiple length group columns.}

\item{line_size}{Numeric. The thickness of the line in the plot. Default is 1.5.}

\item{line_color}{Character. The color of the line in the plot. Default is "SBLack".}

\item{line_type}{Character. The type of the line in the plot. Default is "solid".#' @return A ggplot object representing the plot.}

\item{n_col}{Integer, number of columns in facet_wrap.}

\item{scales}{Character, scales for facet_wrap.}
}
\description{
This function takes a list and a data frame as input, and plots SBL (counts in length groups) over the years using ggplot2. The plot is faceted by the length groups.
}
\examples{
\dontrun{
plot_SBL(model_result, data.CatL)
}
}
