% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SSB_Rec.R
\name{plot_SSB_Rec}
\alias{plot_SSB_Rec}
\title{Plot the Relationship between Spawning Stock Biomass (SSB) and Recruitment (Rec)}
\usage{
plot_SSB_Rec(
  model_result,
  point_size = 2,
  point_color = "black",
  point_shape = 16,
  return_data = FALSE
)
}
\arguments{
\item{model_result}{A list containing the model results. It includes a component "report" that comprises "SSB" and "Rec".}

\item{point_size}{Numeric. The size of the points in the scatter plot. Default is 2.}

\item{point_color}{Character. The color of the points in the scatter plot. Default is "black".}

\item{point_shape}{Numeric. The shape of the points in the scatter plot, as an integer value (see ?points in base R for more info). Default is 16 (filled circle).}

\item{return_data}{A logical indicating whether to return the processed data alongside the plot. Default is FALSE.}
}
\value{
A ggplot object representing the scatter plot of SSB versus Recruitment.
}
\description{
This function takes the model result as an input and creates a scatter plot to visualize the relationship between
Spawning Stock Biomass (SSB) and Recruitment (Rec). These metrics are fundamental in the context of fishery stock assessment.
}
\examples{
\dontrun{
# The model_result is typically obtained from the run_acl() function. Here, we create a dummy example.
# Use 'run_acl' to get 'model_result'
model_result <- run_acl(...)

# Call the function
p <- plot_SSB_Rec(model_result)

# Print the plot
print(p)
}
}
