% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_yearly_length_distributions.R
\name{plot_yearly_length_distributions}
\alias{plot_yearly_length_distributions}
\title{Plot yearly length distributions}
\usage{
plot_yearly_length_distributions(
  model_result,
  point_size = 1.5,
  point_color = "black",
  point_shape = 1,
  line_size = 1.2,
  line_color = "black",
  line_type = "solid",
  facet_ncol = 5,
  facet_scales = "fixed"
)
}
\arguments{
\item{model_result}{A list that contains model output.}

\item{point_size}{Character. The size of the point in the plot. Default is "solid".#' @return A ggplot object representing the plot.}

\item{point_color}{Character. The color of the point in the plot. Default is "black".}

\item{point_shape}{Character. The shape of the point in the plot. Default is "solid".#' @return A ggplot object representing the plot.}

\item{line_size}{Numeric. The thickness of the line in the plot. Default is 1.5.}

\item{line_color}{Character. The color of the line in the plot. Default is "black".}

\item{line_type}{Character. The type of the line in the plot. Default is "solid".#' @return A ggplot object representing the plot.}

\item{facet_ncol}{Integer, number of columns in facet_wrap.}

\item{facet_scales}{Character, scales for facet_wrap.
The list should have a "report" component which contains "Elog_index" component,
and also "len_mid" and "year" components.}
}
\value{
A ggplot object representing the plot.
}
\description{
This function takes a model_result list as input, extracts the Elog_index, len_mid, and year,
and then plots the yearly length distributions using ggplot2.
}
\examples{
\dontrun{
plot_yearly_length_distributions(model_result)
}
}
