% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_alscl1.R
\name{run_alscl1}
\alias{run_alscl1}
\title{Run stock assessment model for user's data}
\usage{
run_alscl1(
  data.CatL,
  data.wgt,
  data.mat,
  rec.age,
  nage,
  M,
  sel_L50,
  sel_L95,
  parameters = NULL,
  parameters.L = NULL,
  parameters.U = NULL,
  map = NULL
)
}
\arguments{
\item{data.CatL}{A matrix containing the length grouping of the catch length.}

\item{data.wgt}{A matrix containing the length grouping of the weight.}

\item{data.mat}{A matrix containing the length grouping of the maturity.}

\item{rec.age}{Numeric, the age of recruitment.}

\item{nage}{Numeric, number of age classes.}

\item{M}{Numeric, natural mortality (default: 0.2)}

\item{sel_L50}{Numeric, length at which 50\% of individuals are mature.}

\item{sel_L95}{Numeric, length at which 95\% of individuals are mature.}

\item{parameters}{A list containing the custom initial values for the parameters (default is NULL).}

\item{parameters.L}{A list containing the custom lower bounds for the parameters (default is NULL).}

\item{parameters.U}{A list containing the custom upper bounds for the parameters (default is NULL).}

\item{map}{A list containing the custom values for the map elements (default is NULL).}
}
\value{
A list containing the results of the stock assessment model.
}
\description{
This function runs a stock assessment model using the specified iteration range.
}
