% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stock_assessment.R
\name{run_stock_assessment}
\alias{run_stock_assessment}
\title{Run stock assessment model}
\usage{
run_stock_assessment(
  iter_range = 4:100,
  sim_data_path = ".",
  output_dir = ".",
  parameters = NULL,
  parameters.L = NULL,
  parameters.U = NULL,
  map = NULL,
  M = 0.2
)
}
\arguments{
\item{iter_range}{A numeric vector specifying the range of iterations to run the stock assessment model for (default is 4:100).}

\item{sim_data_path}{A character string specifying the path to the folder containing the simulation data files (default is the current working directory).}

\item{output_dir}{Character, the directory where simulation results will be saved.(default is the current working directory).}

\item{parameters}{A list containing the custom initial values for the parameters (default is NULL).}

\item{parameters.L}{A list containing the custom lower bounds for the parameters (default is NULL).}

\item{parameters.U}{A list containing the custom upper bounds for the parameters (default is NULL).}

\item{map}{A list containing the custom values for the map elements (default is NULL).}

\item{M}{Numeric, natural mortality (default: 0.2)}
}
\value{
A list containing the results of the stock assessment model.
}
\description{
This function runs a stock assessment model using the specified iteration range.
}
