% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cal.R
\name{sim_cal}
\alias{sim_cal}
\title{Calculate biological and fishing variables}
\usage{
sim_cal(params)
}
\arguments{
\item{params}{A list containing the parameters variables.}
}
\value{
A list containing the biological and fishing variables.
}
\description{
This function calculates and returns biological and fishing variables for the simulation model.
}
