% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate Fishery Dynamics}
\usage{
simulate_data(bio_fishing_vars, init_state_vars, nyear, output_dir = ".")
}
\arguments{
\item{bio_fishing_vars}{A list containing the biological and fishing variables.}

\item{init_state_vars}{A list containing the initial state variables.}

\item{nyear}{Integer, number of years to simulate.}

\item{output_dir}{Character, the directory where simulation results will be saved.(default is the current working directory).}
}
\value{
A list containing simulated fishery data.
}
\description{
This function simulates fishery dynamics based on the given biological and fishing variables and initial state variables.
}
