% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_func.R
\name{VB_func}
\alias{VB_func}
\title{Von Bertalanffy growth function}
\usage{
VB_func(Linf, k, t0, age)
}
\arguments{
\item{Linf}{The asymptotic length.}

\item{k}{The growth coefficient.}

\item{t0}{The theoretical age at zero length.}

\item{age}{Age of the fish.}
}
\value{
Length at age.
}
\description{
Calculate length at age using the von Bertalanffy growth function.
}
