% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_model.R
\name{diagnose_model}
\alias{diagnose_model}
\title{Diagnose model results}
\usage{
diagnose_model(data.CatL, model_result)
}
\arguments{
\item{data.CatL}{A matrix containing the length grouping of the catch length,
which represents the observed catch data in different length groups across years.}

\item{model_result}{A list that contains the model output. The list should have a "report" component which contains an "Elog_index" component representing counts in length groups.}
}
\value{
NULL. The function prints out the diagnostics and does not return anything.
}
\description{
This function receives a model output and performs various checks, including checking if the model has reached its boundaries, if it has converged, and printing the final likelihood estimate.
}
\examples{
\dontrun{
diagnose_model(data.CatL,model_result)
}
}
