% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abundance.R
\name{plot_abundance}
\alias{plot_abundance}
\title{Plot number over the years}
\usage{
plot_abundance(
  model_result,
  line_size = 1.2,
  line_color = "red",
  line_type = "solid",
  se = FALSE,
  se_color = "red",
  se_alpha = 0.2,
  type = c("N", "NA"),
  facet_ncol = 3,
  facet_scales = "free"
)
}
\arguments{
\item{model_result}{A list that contains model output. The list should have a "report" component which contains a "N" component representing number.}

\item{line_size}{Numeric. The thickness of the line in the plot. Default is 1.2.}

\item{line_color}{Character. The color of the line in the plot. Default is "black".}

\item{line_type}{Character. The type of the line in the plot. Default is "solid".}

\item{se}{Logical. Whether to calculate and plot standard error as confidence intervals. Default is FALSE.}

\item{se_color}{Character. The color of the confidence interval ribbon. Default is "blue".}

\item{se_alpha}{Numeric. The transparency of the confidence interval ribbon. Default is 0.2.}

\item{type}{Character. It specifies which the abundance is ploted for "N" or "NA". Default is "N".}

\item{facet_ncol}{Number of columns in facet wrap. Default is 3.}

\item{facet_scales}{Scales for facet wrap. Default is "free".}
}
\value{
A ggplot object representing the plot.
}
\description{
This function takes a list and a data frame as input, and plots number over the years using ggplot2.
}
\examples{
\dontrun{
plot_abundance(model_result, line_size = 1.2, line_color = "red", line_type = "solid")
}
}
