% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomass.R
\name{plot_biomass}
\alias{plot_biomass}
\title{Plot biomass over the years}
\usage{
plot_biomass(
  model_result,
  data.CatL,
  line_size = 1.2,
  line_color = "black",
  line_type = "solid"
)
}
\arguments{
\item{model_result}{A list that contains model output. The list should have a "report" component which contains a "B" component representing biomass.}

\item{data.CatL}{A data frame that contains a "Year" column.}

\item{line_size}{Numeric. The thickness of the line in the plot. Default is 1.5.}

\item{line_color}{Character. The color of the line in the plot. Default is "black".}

\item{line_type}{Character. The type of the line in the plot. Default is "solid".}
}
\value{
A ggplot object representing the plot.
}
\description{
This function takes a list and a data frame as input, and plots biomass over the years using ggplot2.
}
\examples{
\dontrun{
plot_biomass(model_result, data.CatL, line_size = 1.2, line_color = "red", line_type = "solid")
}
}
