% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomass.R
\name{plot_biomass}
\alias{plot_biomass}
\title{Plot biomass over the years}
\usage{
plot_biomass(
  model_result,
  line_size = 1.2,
  line_color = "red",
  line_type = "solid",
  se = FALSE,
  se_color = "red",
  se_alpha = 0.2,
  type = c("B", "BL"),
  facet_ncol = 3,
  facet_scales = "free"
)
}
\arguments{
\item{model_result}{A list that contains model output. The list should have a "report" component which contains a "B" component representing biomass.}

\item{line_size}{Numeric. The thickness of the line in the plot. Default is 1.2.}

\item{line_color}{Character. The color of the line in the plot. Default is "black".}

\item{line_type}{Character. The type of the line in the plot. Default is "solid".}

\item{se}{Logical. Whether to calculate and plot standard error as confidence intervals. Default is FALSE.}

\item{se_color}{Character. The color of the confidence interval ribbon. Default is "blue".}

\item{se_alpha}{Numeric. The transparency of the confidence interval ribbon. Default is 0.2.}

\item{type}{Character. It specifies which the biomass is plot for "B" , "BL" . Default is "N".}

\item{facet_ncol}{Integer, number of columns in facet_wrap.}

\item{facet_scales}{Character, scales for facet_wrap.}
}
\value{
A ggplot object representing the plot.
}
\description{
This function takes a list and a data frame as input, and plots biomass over the years using ggplot2.
}
\examples{
\dontrun{
plot_biomass(model_result, line_size = 1.2, line_color = "red", line_type = "solid")
}
}
