% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deviacne.R
\name{plot_deviance}
\alias{plot_deviance}
\title{Plot Deviation of Recruitment (R) and Fishing Mortality (F) Over Years}
\usage{
plot_deviance(
  model_result,
  se = TRUE,
  point_size = 3,
  point_color = "white",
  point_shape = 21,
  line_size = 1,
  line_color = "black",
  line_type = "solid",
  se_color = "black",
  se_width = 0.5,
  facet_ncol = 3,
  facet_scales = "free",
  log = T,
  type = c("R", "F")
)
}
\arguments{
\item{model_result}{A list that contains the Annual Catch Limit (ACL) model output.}

\item{se}{Logical, whether to draw the standard error bar.}

\item{point_size}{Character. The size of the point in the plot. Default is "solid".}

\item{point_color}{Character. The color of the point in the plot. Default is "white".}

\item{point_shape}{Character. The shape of the point in the plot. Default is "solid".}

\item{line_size}{Numeric, the line size.}

\item{line_color}{Character, the line color.}

\item{line_type}{Character, the line type.}

\item{se_color}{Character, the color of the standard error.}

\item{se_width}{Numeric, the width of the standard error.}

\item{facet_ncol}{Number of columns in facet wrap. Default is 3.}

\item{facet_scales}{Scales for facet wrap. Default is "free".}

\item{log}{Logical, whether to apply an exponential transformation to the results.}

\item{type}{Character. It specifies whether the deviation plot is for "R" or "F". Default is "R".}
}
\value{
A ggplot2 object.
}
\description{
This function calculates the deviations of R and F in the ACL model results,
which are used to measure the deviance of the model. These deviations are calculated
as the difference between the observed and expected values, with the model's residuals
as the measurement errors. This function can plot either R or F deviation over years,
depending on the 'type' parameter.
}
\examples{
\dontrun{
# If you want to plot the deviation of recruitment (R) with standard error bars
plot_deviance(model_result = your_model_results, se = TRUE, type = "R",
              point_size = 1, point_color = "black", point_shape = 1,
              line_size = 1, line_color = "black", line_type = "solid",
              se_color = "blue", se_alpha = 0.4)

# If you do not want to plot the standard error bars
plot_deviance(model_result = your_model_results, se = FALSE, type = "R",
              point_size = 1, point_color = "black", point_shape = 1,
              line_size = 1, line_color = "black", line_type = "solid")

# If you want to plot the deviation of fishing mortality (F) with standard error bars
plot_deviance(model_result = your_model_results, se = TRUE, type = "F",
              point_size = 1, point_color = "black", point_shape = 1,
              line_size = 1, line_color = "black", line_type = "solid",
              se_color = "blue", se_alpha = 0.4)
}
}
