% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fishing_mortality.R
\name{plot_fishing_mortality}
\alias{plot_fishing_mortality}
\title{Plot Fishing Mortality Over Years from an Age-Structured Assessment Model (ACL)}
\usage{
plot_fishing_mortality(
  model_result,
  line_size = 1,
  line_color = "red",
  line_type = "solid",
  facet_ncol = 3,
  facet_scales = "free",
  se = FALSE,
  se_color = "red",
  se_alpha = 0.2
)
}
\arguments{
\item{model_result}{A list obtained from the \code{run_acl} function. This list should contain a "report" component that includes
a "F" component representing fishing mortality for each age group over years.}

\item{line_size}{Numeric. The thickness of the line in the plot. Default is 1.}

\item{line_color}{Character. The color of the line in the plot. Default is "red".}

\item{line_type}{Character. The type of the line in the plot. Default is "solid".}

\item{facet_ncol}{Integer. The number of columns in facet_wrap.}

\item{facet_scales}{Character. Scales for facet_wrap. Default is "free".}

\item{se}{Logical. Whether to calculate and plot standard error as confidence intervals. Default is FALSE. If TRUE, standard error
will be calculated and confidence intervals will be shown as a shaded area around the line.}

\item{se_color}{Character. The color of the confidence interval ribbon. Default is "red".}

\item{se_alpha}{Numeric. The transparency of the confidence interval ribbon. Default is 0.2.}
}
\value{
A ggplot object representing the plot of fishing mortality over years.
}
\description{
This function takes the output from the \code{run_acl} function and plots fishing mortality over the years using ggplot2.
The function can also calculate and display the standard error as confidence intervals.
}
\examples{
\dontrun{
# Use 'run_acl' to get 'model_result'
model_result <- run_acl(...)

# Call the function with standard error set to TRUE
p_mortality <- plot_fishing_mortality(model_result, se = TRUE)

# Print the plot with standard error
print(p_mortality)

# Call the function with standard error set to FALSE
p_mortality <- plot_fishing_mortality(model_result, se = FALSE)

# Print the plot without standard error
print(p_mortality)
}
}
