% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot Recruitment over the years}
\usage{
plot_recruitment(
  model_result,
  line_size = 1.2,
  line_color = "red",
  line_type = "solid",
  se = FALSE,
  se_color = "red",
  se_alpha = 0.2,
  return_data = FALSE
)
}
\arguments{
\item{model_result}{A list that contains the model output. The list should have a "report" component which contains a "Rec" component representing Recruitment.}

\item{line_size}{Numeric. Specifies the thickness of the line in the plot. Default is 1.2.}

\item{line_color}{Character. Specifies the color of the line in the plot. Default is "red".}

\item{line_type}{Character. Specifies the type of the line in the plot. Default is "solid".}

\item{se}{Logical. Determines whether to calculate and plot the standard error as confidence intervals. Default is FALSE.}

\item{se_color}{Character. Specifies the color of the confidence interval ribbon. Default is "red".}

\item{se_alpha}{Numeric. Specifies the transparency of the confidence interval ribbon. Default is 0.2.}

\item{return_data}{A logical indicating whether to return the processed data alongside the plot. Default is FALSE.}
}
\value{
A ggplot object representing the plot.
}
\description{
This function visualizes the recruitment of a population over a period of time.
Recruitment is a key parameter in population dynamics, representing the amount of new individuals added to a population,
either by birth or immigration, over a specific time frame. The resulting plot is a line graph, with the option to add a
ribbon that illustrates the standard error, if the \code{se} argument is set to TRUE.
}
\examples{
\dontrun{
# Use 'run_acl' to obtain 'model_result'
model_result <- run_acl(...)

# Plot recruitment without standard error
plot_recruitment(model_result)

# Plot recruitment with standard error
plot_recruitment(model_result, se = TRUE)
}

}
