% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot model residuals with facets}
\usage{
plot_residuals(
  model_result,
  f = 0.4,
  line_color = "black",
  smooth_color = "blue",
  hline_color = "red",
  line_size = 1,
  facet_scales = "free",
  type = c("length", "age")
)
}
\arguments{
\item{model_result}{A list that contains the model output. The list should have a "report" component which contains a "resid_index" component representing the residuals for each index.}

\item{f}{Numeric. The smoother span. This gives the proportion of points in the plot which influence the smooth at each value. Larger values give more smoothness. Default is 0.4.}

\item{line_color}{Character. The color of the lines in the plot. Default is "black".}

\item{smooth_color}{Character. The color of the smooth line in the plot. Default is "blue".}

\item{hline_color}{Character. The color of the horizontal line in the plot. Default is "red".}

\item{line_size}{Numeric. The size of the lines in the plot. Default is 1.}

\item{facet_scales}{Character. The scale argument for facet_wrap. Default is "free".}

\item{type}{Character. It specifies whether the residuals are calculated for "length" or "age". Default is "length".}
}
\value{
A ggplot object representing the facet plot.
}
\description{
This function creates a facet plot of the residuals for each index in the model result. Each index is represented in a separate facet.
}
\examples{
\dontrun{
plot_residuals(model_result)
}
}
