% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot the Distribution of Model Residuals from an Age-Structured Assessment Model (ACL)}
\usage{
plot_residuals(
  model_result,
  f = 0.4,
  line_color = "black",
  smooth_color = "blue",
  hline_color = "red",
  line_size = 1,
  facet_scales = "free",
  facet_ncol = NULL,
  type = c("length", "year")
)
}
\arguments{
\item{model_result}{A list obtained from the \code{run_acl} function. This list should contain a "report" component that includes
a "resid_index" component, representing the residuals for each index.}

\item{f}{Numeric. The smoother span for the loess smooth line in the plot. This gives the proportion of points in the plot which
influence the smooth at each value. Larger values result in more smoothing. Default is 0.4.}

\item{line_color}{Character. The color of the line in the plot. Default is "black".}

\item{smooth_color}{Character. The color of the smooth line in the plot. Default is "blue".}

\item{hline_color}{Character. The color of the horizontal line (at y=0) in the plot. Default is "red".}

\item{line_size}{Numeric. The size of the lines in the plot. Default is 1.}

\item{facet_scales}{Character. The "scales" argument for the facet_wrap function in ggplot2. Default is "free".}

\item{facet_ncol}{Numeric, optional. Number of columns in facet wrap.  Default is NULL.}

\item{type}{Character. Specifies whether the residuals are calculated over "length" or "year". Default is "length".}
}
\value{
A ggplot object representing the facet plot of residuals.
}
\description{
This function takes the output from the \code{run_acl} function and creates a facet plot of the residuals for each index.
The residuals reflect the difference between the observed and expected values from the ACL model, a type of age-structured
assessment model commonly used in fishery science. The function can handle two types of residuals:
\enumerate{
\item "length", where residuals are calculated over length groups for each year, and
\item "year", where residuals are calculated over years for each length group.
Each index is represented in a separate facet.
}
}
\examples{
\dontrun{
# Use 'run_acl' to get 'model_result'
model_result <- run_acl(...)

# Call the function with type = "length"
p_length <- plot_residuals(model_result, type = "length")

# Print the plot for "length"
print(p_length)

# Call the function with type = "year"
p_year <- plot_residuals(model_result, type = "year")

# Print the plot for "year"
print(p_year)
}
}
