% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ridges.R
\name{plot_ridges}
\alias{plot_ridges}
\title{Plot ridges of Observed and Estimated Catch-at-Length Over Years}
\usage{
plot_ridges(model_result, ridges_alpha = 0.8)
}
\arguments{
\item{model_result}{A list containing the model result, which includes a data frame of Elog_index and logN_at_len, len_mid, and year.}

\item{ridges_alpha}{Numeric. Specifies the transparency of the ridgeline plots. Default is 0.8.}
}
\value{
A grid arranged ggplot object.
}
\description{
This function takes the model result as an input and creates two ridgeline plots to
visualize the density distribution of observed and estimated log catch at length over different years.
The density distribution is calculated by transforming the abundance data with the exponential function.
}
\examples{
\dontrun{
# Assume model_result is the result obtained from function 'run_acl'.
p <- plot_ridges(model_result)
print(p)
}
}
