% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SSB.R
\name{plot_SSB}
\alias{plot_SSB}
\title{Plot Spawning Stock Biomass (SSB) or Spawning Biomass at Length (SBL) Over the Years}
\usage{
plot_SSB(
  model_result,
  line_size = 1.2,
  line_color = "red",
  line_type = "solid",
  se = FALSE,
  se_color = "red",
  se_alpha = 0.2,
  type = c("SSB", "SBL"),
  facet_ncol = NULL,
  facet_scales = "free",
  return_data = FALSE
)
}
\arguments{
\item{model_result}{A list that contains model output. The list should have a "report" component which contains a "SSB" or "SBL" component representing Spawning Stock Biomass or Spawning Biomass at Length respectively.}

\item{line_size}{Numeric, optional. The thickness of the line in the plot. Default is 1.2.}

\item{line_color}{Character, optional. The color of the line in the plot. Default is "red".}

\item{line_type}{Character, optional. The type of the line in the plot. Default is "solid".}

\item{se}{Logical, optional. Whether to calculate and plot standard error as confidence intervals. Default is FALSE.}

\item{se_color}{Character, optional. The color of the confidence interval ribbon. Default is "red".}

\item{se_alpha}{Numeric, optional. The transparency of the confidence interval ribbon. Default is 0.2.}

\item{type}{Character, optional. It specifies whether to plot for "SSB" or "SBL". Default is "SSB".}

\item{facet_ncol}{Numeric, optional. Number of columns in facet wrap. This parameter is only applicable when type is "SBL". Default is NULL.}

\item{facet_scales}{Character, optional. Scales for facet wrap. This parameter is only applicable when type is "SBL". Default is "free".}

\item{return_data}{A logical indicating whether to return the processed data alongside the plot. Default is FALSE.}
}
\value{
A ggplot object representing the plot of Spawning Stock Biomass or Spawning Biomass at Length over the years.
}
\description{
This function takes a list and a data frame as input, and plots SSB or SBL over the years using ggplot2.
}
\examples{
\dontrun{
# Use 'run_acl' to get 'model_result'
model_result <- run_acl(...)


# Plot SSB with confidence intervals
plot_SSB(model_result = model_result, type = "SSB", se = TRUE)

# Plot SSB without confidence intervals
plot_SSB(model_result = model_result, type = "SSB", se = FALSE)

# Plot SBL with confidence intervals
plot_SSB(model_result = model_result, type = "SBL", se = TRUE)

# Plot SBL without confidence intervals
plot_SSB(model_result = model_result, type = "SBL", se = FALSE)
}
}
