% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SSB.R
\name{plot_SSB}
\alias{plot_SSB}
\title{Plot Spawning stock biomass over the years}
\usage{
plot_SSB(
  model_result,
  line_size = 1.2,
  line_color = "red",
  line_type = "solid",
  se = FALSE,
  se_color = "red",
  se_alpha = 0.2
)
}
\arguments{
\item{model_result}{A list that contains model output. The list should have a "report" component which contains a "SSB" component representing SSB}

\item{line_size}{Numeric. The thickness of the line in the plot. Default is 1.2.}

\item{line_color}{Character. The color of the line in the plot. Default is "black".}

\item{line_type}{Character. The type of the line in the plot. Default is "solid".}

\item{se}{Logical. Whether to calculate and plot standard error as confidence intervals. Default is FALSE.}

\item{se_color}{Character. The color of the confidence interval ribbon. Default is "blue".}

\item{se_alpha}{Numeric. The transparency of the confidence interval ribbon. Default is 0.2.}
}
\value{
A ggplot object representing the plot.
}
\description{
This function takes a list and a data frame as input, and plots SSB over the years using ggplot2.
}
\examples{
\dontrun{
plot_SSB(model_result, line_size = 1.2, line_color = "red", line_type = "solid")
}
}
