% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SSB_Rec.R
\name{plot_SSB_Rec}
\alias{plot_SSB_Rec}
\title{Plot SSB and Rec from model results}
\usage{
plot_SSB_Rec(
  model_result,
  point_size = 2,
  point_color = "black",
  point_shape = 16
)
}
\arguments{
\item{model_result}{A list containing model results, including "report" that has "SSB" and "Rec".}

\item{point_size}{Character. The size of the point in the plot. Default is "solid".#' @return A ggplot object representing the plot.}

\item{point_color}{Character. The color of the point in the plot. Default is "black".}

\item{point_shape}{Character. The shape of the point in the plot. Default is "solid".#' @return A ggplot object representing the plot.}
}
\value{
ggplot object
}
\description{
This function takes the model result, extracts the SSB and Rec components, and plots them.
}
\examples{
\dontrun{
plot_SSB_Rec(model_result)
}
}
