% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_acl.R
\name{retro_acl}
\alias{retro_acl}
\title{Retrospective analysis function with plot option}
\usage{
retro_acl(
  nyear,
  data.CatL,
  data.wgt,
  data.mat,
  rec.age,
  nage,
  M,
  sel_L50,
  sel_L95,
  parameters = NULL,
  parameters.L = NULL,
  parameters.U = NULL,
  map = NULL,
  len_mid = NULL,
  len_border = NULL,
  plot = FALSE,
  line_size = 1.2,
  point_size = 3,
  point_shape = 21,
  facet_scales = "free",
  facet_col = NULL,
  facet_row = NULL
)
}
\arguments{
\item{nyear}{Number of years to look back in the retrospective analysis.}

\item{data.CatL}{Data for CatL.}

\item{data.wgt}{Data for weight.}

\item{data.mat}{Data for mat.}

\item{rec.age}{Recreational age.}

\item{nage}{Number of age.}

\item{M}{Natural mortality rate.}

\item{sel_L50}{Length at 50\% selection.}

\item{sel_L95}{Length at 95\% selection.}

\item{parameters}{Optional parameters for the ACL model.}

\item{parameters.L}{Optional L parameters for the ACL model.}

\item{parameters.U}{Optional U parameters for the ACL model.}

\item{map}{Optional map for the ACL model.}

\item{len_mid}{Optional length mid for the ACL model.}

\item{len_border}{Optional length border for the ACL model.}

\item{plot}{Logical indicating whether to plot the results. Defaults to FALSE.}

\item{line_size}{Line size for the plot. Defaults to 1.}

\item{point_size}{Character. The size of the point in the plot. Default is "solid".}

\item{point_shape}{Character. The shape of the point in the plot. Default is "solid".}

\item{facet_scales}{Argument passed to \code{facet_wrap()}. Defaults to "free".}

\item{facet_col}{Defaults to "NULL".}

\item{facet_row}{Defaults to "NULL".}
}
\value{
If \code{plot} is TRUE, returns a ggplot object. Otherwise, returns a dataframe with the results.
}
\description{
Retrospective analysis function with plot option
}
\examples{
\dontrun{
retro_acl(nyear, data.CatL, data.wgt, data.mat, rec.age, nage, M, sel_L50, sel_L95, plot = TRUE)
}
}
