% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_acl.R
\name{retro_acl}
\alias{retro_acl}
\title{Retrospective Analysis in Annual Catch Limit (ACL) with Plotting Option}
\usage{
retro_acl(
  nyear,
  data.CatL,
  data.wgt,
  data.mat,
  rec.age,
  nage,
  M,
  sel_L50,
  sel_L95,
  parameters = NULL,
  parameters.L = NULL,
  parameters.U = NULL,
  map = NULL,
  len_mid = NULL,
  len_border = NULL,
  plot = FALSE,
  line_size = 1.2,
  point_size = 3,
  point_shape = 21,
  facet_scales = "free",
  facet_col = NULL,
  facet_row = NULL,
  train_times = 1
)
}
\arguments{
\item{nyear}{The number of years to look back in the retrospective analysis.}

\item{data.CatL}{The Catch-at-Length data.}

\item{data.wgt}{The weight data.}

\item{data.mat}{The maturity data.}

\item{rec.age}{The recreational age.}

\item{nage}{The number of age.}

\item{M}{The natural mortality rate.}

\item{sel_L50}{The length at 50\% selection.}

\item{sel_L95}{The length at 95\% selection.}

\item{parameters}{Optional parameters for the ACL model.}

\item{parameters.L}{Optional lower limit parameters for the ACL model.}

\item{parameters.U}{Optional upper limit parameters for the ACL model.}

\item{map}{Optional map for the ACL model.}

\item{len_mid}{Optional length mid for the ACL model.}

\item{len_border}{Optional length border for the ACL model.}

\item{plot}{Logical indicating whether to plot the results. Defaults to FALSE.}

\item{line_size}{The line size for the plot. Defaults to 1.2.}

\item{point_size}{The size of the point in the plot. Defaults to 3.}

\item{point_shape}{The shape of the point in the plot. Default is 21.}

\item{facet_scales}{Argument passed to \code{facet_wrap()}. Defaults to "free".}

\item{facet_col}{The number of columns for facet_wrap. Defaults to NULL.}

\item{facet_row}{The number of rows for facet_wrap. Defaults to NULL.}

\item{train_times}{Numeric, the number of times the model is to be trained,}
}
\value{
If \code{plot} is TRUE, returns a ggplot object and a dataframe with the results. Otherwise, returns only the dataframe with the results.
}
\description{
Conduct a retrospective analysis using an Annual Catch Limit (ACL) model in fisheries stock assessment.
The model calculates Mohn's rho, a measure of retrospective bias in the model, which helps in understanding
potential biases in the ACL calculations and for setting appropriate catch limits.
Mohn's rho is calculated by comparing model parameters from a retrospective analysis with model parameters
from the full data set.
}
\examples{
\dontrun{
# Assume we have the necessary data and parameters for the ACL model
data.CatL <- get_CatL_data()
data.wgt <- get_wgt_data()
data.mat <- get_mat_data()
rec.age <- 2
nage <- 10
M <- 0.2
sel_L50 <- 20
sel_L95 <- 30

# Conduct a retrospective analysis for the past 5 years
retro_acl(nyear = 5, data.CatL, data.wgt, data.mat, rec.age, nage, M, sel_L50, sel_L95, plot = TRUE)
}
}
