% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate Fishery Dynamics}
\usage{
sim_data(bio_fishing_vars, init_state_vars, sim_year = 100, output_dir = ".")
}
\arguments{
\item{bio_fishing_vars}{A list containing the biological and fishing variables.}

\item{init_state_vars}{A list containing the initial state variables.}

\item{sim_year}{Integer, number of years to simulate.Defaults to 100}

\item{output_dir}{Character, the directory where simulation results will be saved.(default is the current working directory).}
}
\value{
A list containing simulated fishery data.
}
\description{
This function simulates fishery dynamics based on the given biological and fishing variables and initial state variables.
}
